Refactor the create method and move to module repo queries
