+++ lib/item_repository.rb
@@ -1,5 +1,6 @@
 require 'csv'
 require_relative '../lib/item'
+require_relative '../lib/merchant'
 require_relative '../lib/modules/repo_queries'
 require 'bigdecimal'
 
@@ -29,23 +30,6 @@ class ItemRepository
     all.find_all do |item|
       range.include?(item.unit_price_to_dollars)
     end
-  end
-
-  def create(attributes)
-    sanitized_attributes = {
-                            name:        attributes[:name],
-                            description: attributes[:description],
-                            unit_price:  attributes[:unit_price],
-                            merchant_id: attributes[:merchant_id]
-    }
-    item = Item.new(sanitized_attributes)
-    item.id = all.max_by do |item|
-      item.id
-    end.id + 1
-    item.created_at = Time.now
-    item.updated_at = Time.now
-    all << item
-    item
   end
 
   def update(id, attributes)
